Lots of changes.
- Fix qualifiers not retain the keyword.
^ this is support via a more generic addition to the
grammar regex system where we can have "verbose AST"
mode that forces (keyword=) to actually emit AST node.

- Plus, I also think this change includes some general
speedup via not pass by value. this is entirely like an
intuitive "oh the feels faster and the numbers look
better maybe". so very much not a definitive win, but
suspected.

- Fix return statements that return a data pack being
parsed as a typed data pack.
^ So what we did was make types parse NOT all keywords
but only the keywords that actually name a type.