Adjust syntax for routed values.
There was an issue where it turns out that variable decl
what a special case of the routed value grammar. We didn't
want such a clash so now doing a re-use of the assignment exp
grammar for routed values.

This makes sense I think. Assignment is an op that takes the value
and places it somewhere. We could consider such routed values
as this kind of placement where the dest is either a struct member
or the function parameter var instance.

Does this actually always mean a runtime op? Well, for compile-time
vars its an op at compile-time. For runtime-vars, such struct or function
calling is a runtime op anyways? Whatever, its OK for now.